% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_updates.R
\name{block_precision_order_2}
\alias{block_precision_order_2}
\title{block_precision_order_2}
\usage{
block_precision_order_2(n)
}
\arguments{
\item{n}{the block size}
}
\value{
The precision matrix K corresponding to a random walk of order 2.
}
\description{
Computes the precision matrix for a random walk of order 2
of the given block size
}
