% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cases_per_time}
\alias{cases_per_time}
\title{Compute expected cases over time}
\usage{
cases_per_time(data, state, smoothed = FALSE, spatial = integer(0))
}
\arguments{
\item{data}{a list containing the data}

\item{state}{the current state of the Markov chain}

\item{smoothed}{set true to smooth the cases (i.e. preclude outbreaks). Defaults to false}

\item{spatial}{the set of spatial regions to compute the expected cases at}
}
\value{
a matrix containing the expected cases in time over the spatial range given
}
\description{
Compute expected cases over time
}
