% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{check_data}
\alias{check_data}
\title{Checks the data and updates it, e.g. for no population where we have cases}
\usage{
check_data(data)
}
\arguments{
\item{data}{the data for the model}
}
\value{
the checked (and possibly updated) data for the model
}
\description{
Checks the data and updates it, e.g. for no population where we have cases
}
