% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_updates.R
\name{construct_block_matrices}
\alias{construct_block_matrices}
\title{construct_block_matrices}
\usage{
construct_block_matrices(block_length)
}
\arguments{
\item{block_length}{the length of the block update}
}
\value{
a list of constructed block matrices containing K_eigen, K_bef and K_aft
}
\description{
Construct the cholesky/eigen matrix for a block update of the given length as well
as pre and post matrices (left and right of main block)
}

