% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{deviance}
\alias{deviance}
\title{Deviance}
\usage{
deviance(data, state)
}
\arguments{
\item{data}{a list containing the data}

\item{state}{the current state of the Markov chain}
}
\value{
a matrix containing the log case rate in time and space
}
\description{
Deviance
}

