% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_updates.R
\name{eigen_scale}
\alias{eigen_scale}
\title{Computes the scaled eigen-vectors of a matrix A}
\usage{
eigen_scale(A)
}
\arguments{
\item{A}{the matrix to compute scaled eigen-vectors on}
}
\value{
the scaled eigenvectors of A
}
\description{
Computes the scaled eigen-vectors of a matrix A
}

