% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{ess}
\alias{ess}
\title{Estimated sample size from MCMC chain}
\usage{
ess(x)
}
\arguments{
\item{x}{A matrix or vector of posterior output for a variable. Rows are samples, columns are chains.}
}
\value{
The expected sample size for the variable (i.e. estimated number of independent samples).
}
\description{
Computation taken from the `rstan` package. Included here so that the package
can be as self-contained as possible.
}
