% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{expected_cases}
\alias{expected_cases}
\title{Computed expected cases per time per space}
\usage{
expected_cases(data, state, smoothed = FALSE)
}
\arguments{
\item{data}{a list containing the data}

\item{state}{the current state of the Markov chain}

\item{smoothed}{set true to smooth the cases (i.e. preclude outbreaks). Defaults to false}
}
\value{
a matrix containing the expected cases in time and space
}
\description{
Computed expected cases per time per space
}

