% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_spatial.R
\name{extract_spatial}
\alias{extract_spatial}
\title{Extract spatial trend(s) from an epiclustR model state}
\usage{
extract_spatial(state, data)
}
\arguments{
\item{state}{the model state}

\item{data}{the data used to fit the model}
}
\value{
The spatial risk adjusted as needed for the differing means through time per period.
pull out U and R from the posterior
}
\description{
Extract spatial trend(s) from an epiclustR model state
}

