% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{fit_chain}
\alias{fit_chain}
\title{Fits an MCMC chain for the epiclustR model for the given dataset}
\usage{
fit_chain(chain, data, prior, control)
}
\arguments{
\item{chain}{the MCMC chain to run}

\item{data}{the data for the model}

\item{prior}{details on the priors for the model}

\item{control}{details on the control for the model}
}
\value{
the posterior for this chain
}
\description{
Fits an MCMC chain for the epiclustR model for the given dataset
}
