% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{fit_model}
\alias{fit_model}
\title{fit_model}
\usage{
fit_model(data, state, prior, control)
}
\arguments{
\item{data}{the data for the model}

\item{state}{the current state of the Markov chain}

\item{prior}{details on the priors for the model}

\item{control}{details on the control for the model}
}
\value{
the posterior for this chain
}
\description{
Fits the epiclustR model for the given dataset
}

