% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_helpers.R
\name{five_colour_map}
\alias{five_colour_map}
\title{Colour regions on a map defined by groupings of spatial units using the five colour algorithm}
\usage{
five_colour_map(mbrg, nb)
}
\arguments{
\item{mbrg}{The mapping from fine spatial units to regions}

\item{nb}{The neighbourhood mapping of spatial units to each other}
}
\value{
A vector containing the numbers 1 through 5 as to which colour each fine spatial unit
should be.
}
\description{
Colour regions on a map defined by groupings of spatial units using the five colour algorithm
}

