% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{init_control}
\alias{init_control}
\title{init_control}
\usage{
init_control(samples = 1000, chains = 4, thinning = 50, burnin = 20,
  parallel = chains > 1, sigmaR = 1, sigmaU = 1, sigmaX = 1,
  blockR = c(4, 5, 9, 11))
}
\arguments{
\item{samples}{The number of posterior samples required. These will be distributed
across the chains. Defaults to 1000.}

\item{chains}{The number of posterior chains to run. Each chain will generate
$samples/chains$ samples. Defaults to 4.}

\item{thinning}{The amount of samples to skip per posterior sample for better
independence between samples, defaults to 50.}

\item{burnin}{The number of posterior samples to discard prior to sampling. Defaults to 20}

\item{parallel}{Whether to use parallel processing. Defaults to TRUE if chains > 1.}

\item{sigmaR}{Standard deviation of Metropolis-Hastings random walk jump for R, defaults to 1.}

\item{sigmaU}{Standard deviation of Metropolis-Hastings random walk jump for U, defaults to 1.}

\item{sigmaX}{Standard deviation of Metropolis-Hastings random walk jump for betaX, defaults to 1.}

\item{blockR}{Block sizes for updating R from the conditional prior distribution. Should be
mutually prime. Defaults to `c(4,5,9,11)`.}
}
\value{
list containing the MCMC control parameters.
}
\description{
Initialize the MCMC control parameters for fitting the epiclustR model
}

