% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_neighbours.R
\name{init_lut}
\alias{init_lut}
\title{Initialize a reverse lookup table from a lookup vector}
\usage{
init_lut(luv)
}
\arguments{
\item{luv}{A lookup vector mapping items to a grouping variable with levels 1 to n}
}
\value{
a list for reverse lookups from levels to vector entries in `luv`
}
\description{
Initialize a reverse lookup table from a lookup vector
}

