% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_neighbours.R
\name{init_region_lut}
\alias{init_region_lut}
\title{Initialize the region to spatial unit lookup table}
\usage{
init_region_lut(mbrg)
}
\arguments{
\item{mbrg}{The spatial unit to region lookup vector}
}
\value{
a list of the spatial units in each region
}
\description{
Initialize the region to spatial unit lookup table
}

