% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{init_state}
\alias{init_state}
\title{Initialize the model state}
\usage{
init_state(ntimes, nspace, nregions, nperiods, fe = -10, pX = 0.1,
  kU = 1, kR = 1)
}
\arguments{
\item{ntimes}{The number of time points}

\item{nspace}{The number of spatial units}

\item{nregions}{The number of regions}

\item{nperiods}{The number of periods}

\item{fe}{The starting value for the fixed effect}

\item{pX}{The starting value for pX, the probability of an outbreak}

\item{kU, kR}{The starting values for the precision of spatial and temporal components.}
}
\value{
A list containing the model state
}
\description{
Initialize the model state
}
