% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_neighbours.R
\name{load_spatial_neighbours}
\alias{load_spatial_neighbours}
\title{Load spatial neighbours data}
\usage{
load_spatial_neighbours(file)
}
\arguments{
\item{file}{the file to load.}
}
\value{
A matrix of spatial neighbours, the first column of which is the number of neighbours
}
\description{
Loads a spatial neighbours file, of the format
}
\details{
0 6
1 5
2 3 4 5 6
2 2 
1 3

where even rows indicate the spatial unit being considered, and
it's number of neighbours, and odd rows indicating what the neighbours are.
i.e. there are 6 spatial units in the
file above, and unit 1 has 5 neighbours which are units 2 through 5,
while unit 2 has 2 neighbours (1 and 3).
}
