% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{log_case_rate}
\alias{log_case_rate}
\title{Computed case rate per time per space}
\usage{
log_case_rate(data, state, smoothed = FALSE, spatial = integer(0))
}
\arguments{
\item{data}{a list containing the data}

\item{state}{the current state of the Markov chain}

\item{smoothed}{set true to smooth the cases (i.e. preclude outbreaks). Defaults to false}

\item{spatial}{the spatial locations to compute the expected cases at.}
}
\value{
a matrix containing the expected cases in time and space
}
\description{
Computed case rate per time per space
}
