% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_region}
\alias{plot_region}
\title{Plot outbreaks from an epiclustR model}
\usage{
plot_region(mod, data, region = NULL, main = region, ...)
}
\arguments{
\item{mod}{the model to plot}

\item{data}{the data for the model (for plotting cases)}

\item{region}{which region to plot. Defaults to NULL (all regions).}

\item{main}{the main title of the plot. Defaults to the region.}

\item{...}{other parameters passed to the plotting functions.}
}
\description{
Plot outbreaks from an epiclustR model
}
