% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_temporal}
\alias{plot_temporal}
\title{Plot the temporal fit of the epiclustR model}
\usage{
plot_temporal(mod, data, spatial = NULL, region = NULL,
  col_cases = "grey80", col_trend = "black", col_outbreaks = "red",
  lwd_cases = 1, lwd_trend = 2, lwd_outbreaks = 1)
}
\arguments{
\item{mod}{the model to plot}

\item{data}{the data for the model (for plotting cases)}

\item{spatial}{the spatial unit(s) to plot the trend for (defaults to NULL, which is all units)}

\item{region}{the region(s) to plot the trend for (defaults to NULL, which is all regions)}

\item{col_cases, col_trend, col_outbreaks}{The colour to use for each component}

\item{lwd_cases, lwd_trend, lwd_outbreaks}{The line width to use for each component}
}
\description{
Plot the temporal fit of the epiclustR model
}
