% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_traces}
\alias{plot_traces}
\title{Plot traces from MCMC chains to assess convergence}
\usage{
plot_traces(mod)
}
\arguments{
\item{mod}{the model to plot}
}
\description{
Plot traces from MCMC chains to assess convergence
}
