% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{table_outbreaks}
\alias{table_outbreaks}
\title{Table of outbreaks from an epiclustR model}
\usage{
table_outbreaks(mod, data, period = NULL, threshold = 0.2)
}
\arguments{
\item{mod}{The model to table}

\item{data}{The data used for the model}

\item{period}{The time period to use. Defaults to NULL which is all time}

\item{threshold}{The threshold for an outbreak. Only outbreaks that exceed this threshold will be displayed. Defaults to 0.2}
}
\description{
Table of outbreaks from an epiclustR model
}

