% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{update}
\alias{update}
\title{Update the MCMC simulation}
\usage{
update(data, i, state, prior, control)
}
\arguments{
\item{data}{a list containing the data}

\item{i}{the current iteration}

\item{state}{the current state of the Markov chain}

\item{prior}{the priors used for the model}

\item{control}{the control parameters for the MCMC algorithm (e.g. tuning, blocks)}
}
\value{
a list containing the updated state
}
\description{
Update the MCMC simulation
}

