% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{reset_acceptance}
\alias{reset_acceptance}
\title{Reset the acceptance variables in the model state}
\usage{
reset_acceptance(state)
}
\arguments{
\item{state}{the current state of the Markov chain}
}
\value{
the state, with the acceptance probabilities reset
}
\description{
Reset the acceptance variables in the model state
}

